ISsue #1057 move fetch lat long to higher level api directory
